Refactor: 等差数列関連クラスを MathEx 配下に移動
- `TAmeAtCoderLibrary/MathEx/Aggregate.cs` ファイルを削除しました。([source: 1])
- `TAmeAtCoderLibrary/MathEx/ArithmeticSequence.cs` ファイルを新規作成し、削除した `Aggregate.cs` の内容を移動しました。([source: 87]-[source: 176])
- 移動に伴い、`ArithmeticSequence` クラスを `MathEx` 静的クラスのネストクラス (`MathEx.ArithmeticSequence`) として再定義しました。([source: 87])

等差数列関連の計算機能は、他の数学関連ユーティリティと共に `MathEx` クラス内にまとめる方が、コードの構造としてより整理され、見通しが良くなると判断したためです。
`Aggregate.cs` というファイル名は、その内容（等差数列の計算）を正確に表していなかったため、`ArithmeticSequence.cs` というより適切なファイル名に変更しました。

- クラス名やメソッド名、機能自体に変更はありません。([source: 3]-[source: 86], [source: 89]-[source: 176])
- 主な変更点は、ファイル名とクラスの配置場所の変更です。
